% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gehan.R
\name{gehan}
\alias{gehan}
\title{Perform Gehan test.}
\usage{
gehan(site, background, alpha, deltaS, power, print = TRUE, plot = TRUE)
}
\arguments{
\item{site}{A dataframe for site measurements. Includes one column of measurements and one column of detect/nondetect designation.}

\item{background}{A dataframe for background measurements. Includes one column of measurements and one column of detect/nondetect designation..}

\item{alpha}{Type I error rate. Options are \code{0.05} (default), \code{0.01}, or \code{0.10}.}

\item{deltaS}{Magnitude of the difference in median site and background concentrations.}

\item{power}{Statistical power. Options are \code{0.75} (default), \code{0.90}, or \code{0.95}.}

\item{print}{Logical. Will display Gehan test results. (default is \code{TRUE})}

\item{plot}{Logical. Will display a histogram of the site and background measurements with other details. (default is \code{FALSE})}
}
\value{
A list containing two objects:
\itemize{
  \item \code{List of parameters and resulting statistics}
     \item \code{alpha} - Type I error rate. 
     \item \code{power} - Statistical power to detect differences. 
     \item \code{deltaS} - Magnitude of the difference in median site and background concentrations.
     \item \code{G} - Gehan test statistic (analogous to a Z statistic). 
     \item \code{p_val} - P-value of the test, computed from a standard normal distribution. 
   \item \code{Data frame used to calculate ranks and scores of each observation}
     \item \code{Samp Measurement} - Measured chemcical concentration of each sample.
     \item \code{Detect (Y/N)} - Inicator measurement detection status where  detect = 0 and nondetect = 1.
     \item \code{Samp Location} - Indicator of location sampled where background = 0 and site = 1.
     \item \code{d} - Rank determining variable 1. 
     \item \code{e} - Rank determining variable 2. 
     \item \code{R} - Calculated rank. 
  \item \code{aR} - Calculated rank score. 
}
}
\description{
Perform Gehan test.
}
\details{
The Gehan test is a nonparametric test that is useful for detecting differences in the median chemical concentrations of 
site and background measurements. It is functionally is similar to the more commonly used Wilcoxon-Rank Sum (WRS) test, but it is 
more powerful in situations where sample size is small, there is a significant number of nondetect measurements, and the measurements 
have different reporting limits. The test evaluates the difference in median site and background chemical concentration by iteratively 
ranking the data by size and detection status. The resulting ranks are used to assign each measurement a score that is used to calculate
the Gehan statistic (G). G follows a standard normal distribution, so a p-value can be computed to assess significance of the test results.
The \code{gehan()} function automatically assesses statistical assumptions of the test based on the user's desired power,
significance level, and sample size, but additional details regarding implementation can be found in 
"Guidance for Environmental Background Analysis Volume III: Groundwater" (2003). 

The function allows users to store the test results as an object. The output is a list containing a list of parameters used in the test, 
the resulting statistics, and the data frame used to calculate the ranks and scores of each observation.
}
\examples{
site_ex1 <- data.frame(samples=c(2, 4, 8, 17, 20, 25, 34, 35, 40, 43), nondetect = c(0, 1, rep(0, 5), 1, rep(0,2)))
bck_ex1 <- data.frame(samples=c(1, 4, 5, 7, 12, 15, 18, 21, 25, 27), nondetect = c(0, 1, 0, 0, 1, 0, 0, 1, 1, 0))
G_result <- gehan(alpha = 0.05, deltaS = 2.0, power = 0.9, site = site_ex1, background = bck_ex1) 

}
\references{
Naval Facilities Engineering Command. (2003, October).  \emph{Guidance for Environmental Background Analysis Volume III: Groundwater.} https://vsp.pnnl.gov/docs/Draft_Guidance_for_Review.pdf.
}
