% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantile 4.9.R
\name{Quantile}
\alias{Quantile}
\title{Perform quantile test.}
\usage{
Quantile(
  site,
  measure.s = site$measurement,
  nd.s = site$nondetect,
  background,
  measure.b = background$measurement,
  nd.b = background$nondetect,
  alpha = 0.05,
  epsilon,
  power = 0.8,
  print = TRUE,
  plot = FALSE
)
}
\arguments{
\item{site}{Dataframe for site measurements.}

\item{background}{Dataframe for background measurements.}

\item{alpha}{Type I error rate. Options are only \code{0.01}, \code{0.025}, \code{0.05}, or \code{0.10}. Must be manually input. (default is \code{0.05})}

\item{epsilon}{A proportion of the site that has concentrations greater than the background. Refer to Table 4.4 of "Guidance for Environmental Background Analysis Volume III: Groundwater" (2003) for possible values.}

\item{power}{Statistical power. Options are only \code{0.80} or \code{0.90}. Must be manually input. (default is \code{0.80})}

\item{print}{Logical. Will display quantile test results. (default is \code{TRUE})}
}
\value{
edit later
}
\description{
Perform quantile test.
}
\examples{
# example code
example_site <- data.frame(samples=c(15, 15, 17, 23, 16, 30, 60, 89, 90, 100), nondetect = c(1,rep(0,9)))
example_background <- data.frame(samples=c(23, 36, 30, 37, 44, 57, 60, 61, 61, 79), nondetect = c(1,1,rep(0,8)))
Quantile(alpha = 0.05, epsilon = 0.5, power = 0.8, site = example_site, background = example_background, plot=T)
}
