% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slippage.R
\name{slippage}
\alias{slippage}
\title{Perform slippage test.}
\usage{
slippage(
  site,
  nd.s = site$nondetect,
  background,
  nd.b = background$nondetect,
  measure.s = site$measurement,
  measure.b = background$measurement,
  epsilon,
  alpha = 0.05,
  power = 0.8,
  print = TRUE,
  plot = FALSE
)
}
\arguments{
\item{site}{A dataframe for site measurements.}

\item{nd.s}{Indicator variable for \code{site} measurements that determines if the corresponding measurement (row) is a nondetect. A value of \code{1} indicates that the row is a nondetect (\code{0} is a detected measurement).}

\item{background}{A dataframe for background measurements.}

\item{nd.b}{Indicator variable for \code{background} measurements that determines if the corresponding measurement (row) is a nondetect. A value of \code{1} indicates that the row is a nondetect (\code{0} is a detected measurement).}

\item{measure.s}{Variable for \code{site} measurements that provides the corresponding concentration.}

\item{measure.b}{Variable for \code{background} measurements that provides the corresponding concentration.}

\item{epsilon}{A number. Refer to Table 4.3 of corresponding literature for possible values.}

\item{alpha}{Type I error rate. Options are only \code{0.05} (default) or \code{0.01}.}

\item{power}{Statistical power. Options are only \code{0.80} (default) or \code{0.90}.}

\item{print}{Logical. Will display slippage test results. (default is \code{TRUE})}

\item{plot}{Logical. Will display a histogram of the site and background measurements with other details. (default is \code{FALSE})}
}
\value{
A list of measures used in the Slippage test, including \code{alpha}, \code{power}, max background measurement, the number of detected site measurements greater than the max background & their corresponding values, and the critical value.
\itemize{
  \item \code{alpha} - Type I error rate.
  \item \code{power} - Statistical power.
  \item \code{maxbg} - Maximum background measurement.
  \item \code{sitemax.n} - The number of detected site measurements greater than the maximum background measurement.
  \item \code{sitemax.m} - The corresponding measurements of the previous component.
  \item \code{critical.value} - The critical value, determined by sample size of site (n) and background (m), including nondetects. Based on Table B-2 and Table B-3 from the literature.
}
}
\description{
Perform slippage test.
}
\details{
The slippage test is one of multiple possible nonparametric tests described in "Guidance for Environmental Background Analysis Volume III: Groundwater" (2003) for determining if a chemical is a Chemical of Potential Concern (COPC).
There are multiple assumptions that need to be upheld during the usage of the \code{slippage()} function.
The maximum background measurement must be a detected measurement; otherwise, an error is generated.
Other conditions are implemented based on the included values in certain tables of the corresponding literature.
\itemize{
\item The sample size of both \code{site} and \code{background} measurements must be between 1 and 50 due to the constraints of Tables B-2 and B-3 in the corresponding literature.
 \item The \code{epsilon} and \code{power} arguments must be able to be located in Table 4-3 of the corresponding literature; otherwise, an error is generated.
 \item  The critical value must be able to be located in Tables B-2 and B-3 of the corresponding literature, otherwise, an error is generated.
}
}
\examples{
set.seed(4321)
example_site <-
  data.frame(measure = round(rchisq(
    n = 50, df = 5, ncp = 6
  ), 1),
  nondetect = rbinom(n = 50, size = 1, p = 0.2))
example_background <-
  data.frame(measure = round(rchisq(n = 50, df = 5), 1),
             nondetect = rbinom(n = 50, size = 1, p = 0.2))
example1 <- slippage(
  site = example_site,
  measure.s = example_site$measure,
  nd.s = example_site$nondetect,
  background = example_background,
  measure.b = example_background$measure,
  nd.b = example_background$nondetect,
  epsilon = 0.6,
  plot = TRUE
)
example1
}
\references{
Naval Facilities Engineering Command. (2003, October).  *Guidance for Environmental Background Analysis Volume III: Groundwater.* https://vsp.pnnl.gov/docs/Draft_Guidance_for_Review.pdf.
}
